requirements.txt
pandas
numpy
pyyaml
matplotlib
tqdm
scipy
alpaca-trade-api
